day 3 ex, still neeed to deplot

+++ contracts/APIConsumer.sol
@@ -4,7 +4,6 @@ import "@chainlink/contracts/src/v0.6/ChainlinkClient.sol";
 import "@chainlink/contracts/src/v0.6/vendor/Ownable.sol";
 
 contract APIConsumer is ChainlinkClient, Ownable {
-
     uint256 public volume;
 
     address private oracle;
@@ -17,7 +16,12 @@ contract APIConsumer is ChainlinkClient, Ownable {
      * Job ID: 29fa9aa13bf1468788b7cc4a500a45b8
      * Fee: 0.1 LINK
      */
-    constructor(address _oracle, string memory _jobId, uint256 _fee, address _link) public {
+    constructor(
+        address _oracle,
+        string memory _jobId,
+        uint256 _fee,
+        address _link
+    ) public {
         if (_link == address(0)) {
             setPublicChainlinkToken();
         } else {
@@ -35,12 +39,18 @@ contract APIConsumer is ChainlinkClient, Ownable {
      * Create a Chainlink request to retrieve API response, find the target
      * data, then multiply by 1000000000000000000 (to remove decimal places from data).
      */
-    function requestVolumeData() public returns (bytes32 requestId)
-    {
-        Chainlink.Request memory request = buildChainlinkRequest(jobId, address(this), this.fulfill.selector);
+    function requestVolumeData() public returns (bytes32 requestId) {
+        Chainlink.Request memory request = buildChainlinkRequest(
+            jobId,
+            address(this),
+            this.fulfill.selector
+        );
 
         // Set the URL to perform the GET request on
-        request.add("get", "https://min-api.cryptocompare.com/data/pricemultifull?fsyms=ETH&tsyms=USD");
+        request.add(
+            "get",
+            "https://min-api.cryptocompare.com/data/pricemultifull?fsyms=ETH&tsyms=USD"
+        );
 
         // Set the path to find the desired data in the API response, where the response format is:
         // {"RAW":
@@ -55,7 +65,7 @@ contract APIConsumer is ChainlinkClient, Ownable {
         request.add("path", "RAW.ETH.USD.VOLUME24HOUR");
 
         // Multiply the result by 1000000000000000000 to remove decimals
-        int timesAmount = 10**18;
+        int256 timesAmount = 10**18;
         request.addInt("times", timesAmount);
 
         // Sends the request
@@ -65,7 +75,9 @@ contract APIConsumer is ChainlinkClient, Ownable {
     /**
      * Receive the response in the form of uint256
      */
-    function fulfill(bytes32 _requestId, uint256 _volume) public recordChainlinkFulfillment(_requestId)
+    function fulfill(bytes32 _requestId, uint256 _volume)
+        public
+        recordChainlinkFulfillment(_requestId)
     {
         volume = _volume;
     }
@@ -75,11 +87,20 @@ contract APIConsumer is ChainlinkClient, Ownable {
      *
      */
     function withdrawLink() external onlyOwner {
-        LinkTokenInterface linkToken = LinkTokenInterface(chainlinkTokenAddress());
-        require(linkToken.transfer(msg.sender, linkToken.balanceOf(address(this))), "Unable to transfer");
+        LinkTokenInterface linkToken = LinkTokenInterface(
+            chainlinkTokenAddress()
+        );
+        require(
+            linkToken.transfer(msg.sender, linkToken.balanceOf(address(this))),
+            "Unable to transfer"
+        );
     }
 
-    function stringToBytes32(string memory source) public pure returns (bytes32 result) {
+    function stringToBytes32(string memory source)
+        public
+        pure
+        returns (bytes32 result)
+    {
         bytes memory tempEmptyStringTest = bytes(source);
         if (tempEmptyStringTest.length == 0) {
             return 0x0;
